refactor: file structure organization (#106)
