+++ jest.config.ts
@@ -25,7 +25,7 @@ const config: Config.InitialOptions = {
 
   // An array of glob patterns indicating a set of files for which coverage information should be collected
   // collectCoverageFrom: ['src/app.controller.ts'],
-  collectCoverageFrom: ['src/logic.service.ts'],
+  collectCoverageFrom: ['src/services/**'],
 
   // The directory where Jest should output its coverage files
   coverageDirectory: '<rootDir>/coverage/',
